backend/requirements.txt
# FastAPI and Server
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LangChain and LangGraph
langchain==0.3.15
langchain-core>=0.3.59
langchain-openai==0.3.17
langchain-chroma==0.2.2
langgraph==0.2.50

# Vector Store
chromadb==0.5.23

# Cohere (for reranking)
cohere==5.13.4

# Google Cloud Storage
google-cloud-storage==2.14.0

# Other utilities
python-dotenv==1.0.1
pyyaml==6.0.2
pandas==2.2.0
openpyxl==3.1.2
